+++ Tests/Integration/MSFT_NetAdapterState.Integration.Tests.ps1
@@ -1,0 +1,72 @@
+$script:DSCModuleName = 'NetworkingDsc'
+$script:DSCResourceName = 'MSFT_NetAdapterState'
+
+Import-Module -Name (Join-Path -Path (Join-Path -Path (Split-Path $PSScriptRoot -Parent) -ChildPath 'TestHelpers') -ChildPath 'CommonTestHelper.psm1') -Global
+
+#region HEADER
+# Integration Test Template Version: 1.1.1
+[System.String] $script:moduleRoot = Split-Path -Parent (Split-Path -Parent $PSScriptRoot)
+if ( (-not (Test-Path -Path (Join-Path -Path $script:moduleRoot -ChildPath 'DSCResource.Tests'))) -or `
+    (-not (Test-Path -Path (Join-Path -Path $script:moduleRoot -ChildPath 'DSCResource.Tests\TestHelper.psm1'))) )
+{
+    & git @('clone', 'https://github.com/PowerShell/DscResource.Tests.git', (Join-Path -Path $script:moduleRoot -ChildPath '\DSCResource.Tests\'))
+}
+
+Import-Module -Name (Join-Path -Path $script:moduleRoot -ChildPath 'DSCResource.Tests\TestHelper.psm1') -Force
+$TestEnvironment = Initialize-TestEnvironment `
+    -DSCModuleName $script:DSCModuleName `
+    -DSCResourceName $script:DSCResourceName `
+    -TestType Integration
+#endregion
+
+# Using try/finally to always cleanup even if something awful happens.
+try
+{
+    # Load the tools required to setup the loopback adapter
+    . (Join-Path -Path (Split-Path -Parent $Script:MyInvocation.MyCommand.Path) -ChildPath 'IntegrationHelper.ps1')
+
+    # Create a Loopback adapter to use to test disabling a NetAdapter
+    $adapterName = 'NetworkingDscLBA'
+    New-IntegrationLoopbackAdapter -AdapterName $adapterName
+
+    #region Integration Tests
+    $ConfigFile = Join-Path -Path $PSScriptRoot -ChildPath "$($script:DSCResourceName).config.ps1"
+    . $ConfigFile -Verbose -ErrorAction Stop
+
+    Describe "$($script:DSCResourceName)_Integration" {
+        It 'Should compile and apply the MOF without throwing' {
+            {
+                & "$($script:DSCResourceName)_Config" -OutputPath $TestDrive
+
+                Start-DscConfiguration `
+                    -Path $TestDrive `
+                    -ComputerName localhost `
+                    -Wait `
+                    -Verbose `
+                    -Force `
+                    -ErrorAction Stop
+            } | Should -Not -Throw
+        }
+
+        It 'Should be able to call Get-DscConfiguration without throwing' {
+            { Get-DscConfiguration -Verbose -ErrorAction Stop } | Should -Not -Throw
+        }
+
+        It 'Should have set the resource and all the parameters should match' {
+            $current = Get-DscConfiguration | Where-Object -FilterScript {
+                $_.ConfigurationName -eq "$($script:DSCResourceName)_Config"
+            }
+            $current.Name     | Should -Be $TestNetAdapterState.Name
+            $current.State    | Should -Be $TestNetAdapterState.State
+        }
+    }
+    #endregion
+}
+finally
+{
+    #region FOOTER
+    Remove-IntegrationLoopbackAdapter -AdapterName $adapterName
+
+    Restore-TestEnvironment -TestEnvironment $TestEnvironment
+    #endregion
+}