+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Fix example publish to PowerShell Gallery by adding `gallery_api`
+  environment variable to `AppVeyor.yml` - fixes [Issue #385](https://github.com/PowerShell/NetworkingDsc/issues/385).
 - MSFT_Proxy:
   - Fixed `ProxyServer`, `ProxyServerExceptions` and `AutoConfigURL`
     parameters so that they correctly support strings longer than 255